Be explicit about cast for d2i_*

+++ ngx_keyless_module.c
@@ -783,7 +783,8 @@ static int ngx_http_keyless_cert_cb(ngx_ssl_conn_t *ssl_conn, void *data)
 
 		p = CBS_data(&child);
 
-		x509 = d2i_X509(NULL, &p, CBS_len(&child));
+		/* A u16 length cannot overflow a long. */
+		x509 = d2i_X509(NULL, &p, (long)CBS_len(&child));
 		if (!x509) {
 			ngx_ssl_error(NGX_LOG_EMERG, c->log, 0, "d2i_X509(...) failed");
 			goto error;
